Task: Rename Cochrane module to SampleSize.
